requirements.txt
ftfy==6.0.3
absl-py==1.2.0
appdirs==1.4.4
aspy.yaml==1.3.0
astor==0.8.1
astroid==2.5.1
attrs==20.3.0
cached-property==1.5.2
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==8.1.3
cycler==0.11.0
decorator==4.4.2
distlib==0.3.1
face-alignment==1.3.5
filelock==3.0.12
gast==0.4.0
grpcio==1.47.0
identify==2.1.0
idna==2.10
imageio==2.20.0
importlib-metadata==4.12.0
iniconfig==1.1.1
ipykernel==4.6.0
ipython==5.3.0
ipython-genutils==0.2.0
isort==5.7.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.4.4
kornia==0.5.1
lazy-object-proxy==1.5.2
llvmlite==0.38.1
lmdb==1.3.0
Markdown==3.4.1
matplotlib==3.4.0
mccabe==0.6.1
networkx==2.8.5
ninja==1.10.2.3
nodeenv==1.5.0
numba==0.55.2
numpy==1.20.1
opencv-python==4.5.5.64
packaging==20.9
pandas==1.4.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pluggy==0.13.1
pre-commit==1.10.4
prompt-toolkit==1.0.18
protobuf==3.15.4
ptyprocess==0.7.0
py==1.10.0
Pygments==2.8.0
pygobject
pylint==2.7.2
pyparsing==2.4.7
pytest==6.2.2
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
pytz==2022.1
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
scikit-image==0.19.3
scipy==1.8.1
simplegeneric==0.8.1
six==1.15.0
ssh-import-id==5.7
tensorboard==2.0.0
tifffile==2022.5.4
timm
toml==0.10.2
torchvision
tornado==6.1
tqdm==4.64.0
traitlets==5.0.5
typing-extensions==4.3.0
unattended-upgrades==0.1
urllib3==1.26.3
virtualenv==20.4.2
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.12.1
zipp==3.8.1
setuptools==60.5.0
regex
easydict
dill
